requirements.txt
pyserial>=3.5
pandas>=1.1.5
matplotlib>=3.3.4
